Lots of tweaks to generated code
Currently, just the classes for each method are generated (roughly in
the style of qtHaskell).  The main difference is that we use type classes
(IsAQPrinter) constraints on self and other qobject-like arguments.
qtHaskell used something a bit more complex with structural phantom
types.

This approach will be much more readable and allow for better
separation of definitions.