Generate enums
Enums are currently defined with the class they are members of.  This
will miss global enums (are there any?).  This will also fail if an
enum from one class is referenced in another.  I do not believe this
happens.  If it is an issue, enums can all be lifted to a common
module (or each to their own modules).

This uses a new class defined in the Smoke module (we don't re-use the
standard Enum class since it implies semantics we don't really want to
provide).  This class is very simple and just converts the ADT to an
Integer value.  This conversion will be used when defining instances
of methods.